Rename components integration tests to resources. Also test in-memory which only has a single version
