+++ tests/integration/integration.go
@@ -22,11 +22,11 @@ import (
 	"github.com/dapr/dapr/tests/integration/framework/binary"
 	"github.com/dapr/dapr/tests/integration/suite"
 
-	_ "github.com/dapr/dapr/tests/integration/suite/components"
 	_ "github.com/dapr/dapr/tests/integration/suite/daprd"
 	_ "github.com/dapr/dapr/tests/integration/suite/healthz"
 	_ "github.com/dapr/dapr/tests/integration/suite/metadata"
 	_ "github.com/dapr/dapr/tests/integration/suite/ports"
+	_ "github.com/dapr/dapr/tests/integration/suite/resources"
 )
 
 func RunIntegrationTests(t *testing.T) {
